Minor update to refresh scripts. Default behavior now is to relaunch old files, only when given an argument (any argument) do they pull new files from GitHub.

+++ refresh_backend.sh
@@ -23,12 +23,16 @@ START_COMMAND="java -jar $JAR_FILENAME"
 #
 #	SCRIPT START
 #
-#	Remove old JAR and .properties file
-rm ${JAR_FILENAME} ${PROPERTIES_FILENAME}
+# If we got an argument (Any argument), pull new files from github
+if [[ $# -gt 0 ]]
+then
+	#	Remove old JAR and .properties file
+	rm ${JAR_FILENAME} ${PROPERTIES_FILENAME}
 
-#	Download new JAR and properties
-wget --no-cache -O ${JAR_FILENAME} ${JAR_LOCATION}
-wget --no-cache -O ${PROPERTIES_FILENAME} ${PROPERTIES_LOCATION}
+	#	Download new JAR and properties
+	wget --no-cache -O ${JAR_FILENAME} ${JAR_LOCATION}
+	wget --no-cache -O ${PROPERTIES_FILENAME} ${PROPERTIES_LOCATION}
+fi
 
 #	Find the Process ID of the old JAR
 OLD_PID=$( pgrep -f "${START_COMMAND}" )
